Add `doas` to allow user switching.
The plan is to switch away from `sudo` completely, however it's still
being installed for backwards-compatibility.

`sudo` is a little unwieldy and hard to configure, e.g. 'preserve env'
doesn't do what we expect (e.g. it doesn't preserve `LD_PRELOAD`, which
is required for PHP's `iconv` extension to work).