Fix crash in the Github Organisation Backup script
It looks like Github doesn't remove old migrations like I initially
thought. This means `octokit.migrations.listForOrg` was attempting to
load _all_ migrations since we started backups.

We hit some sort of tipping point earlier this month where this was just
too much data for the GH server to return and it kept erroring out with
HTTP 502s.

This changes the migration lister to only check the last 10 migrations,
with the assumption that migrations before that have already been
downloaded (since we run this daily).

This also takes the opportunity to upgrade a few dependencies, including
in the `awscli` container (which was failing its build).